Only expose `(|>)` when using OCaml pre-4.01
OCaml 4.01 introduces the `(|>)` operator in `Pervasives`. As a result,
when `open`ing the `QuickCheck_util` module, a warning is given because
a name is aliased.

This in turn causes a compiler error since warnings-as-errors is
enabled.

This commit uses `optcomp` so `(|>)` is not part of the module when
compiled using OCaml >= 4.01.